<L2S20-2> Unit tests and some documentation #time 1h Work logged

+++ eapli.base/eapli.base/base.core/src/main/java/eapli/base/productmanagement/domain/ProductBuilder.java
@@ -24,7 +24,7 @@ public class ProductBuilder implements DomainFactory<Product> {
     private Cash pricePreTax;
     private Cash pricePosTax;
     private BarCode barCode;
-    private double productionCode;
+    private int productionCode;
     private Set<Photo> photosCollection;
 
     public ProductBuilder ofType(ProductCategory productCategory) {
